docker/local-single/services/monitor/requirements.txt
PyAudio==0.2.14
celery==5.4.0
numpy==2.0.1
redis==5.0.7
python-dotenv==1.0.1
pyTelegramBotAPI==4.21.0
boto3==1.34.148
matplotlib==3.9.1
